refs #cursor - hw11. fixes

+++ HW06/src/main/java/hw06/task1/repositories/ProductRepository.java
@@ -5,6 +5,8 @@ import hw06.task1.entities.Product;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
+import java.math.BigDecimal;
+
 /**
  * @author YevhenKovalevskyi
  */
@@ -12,7 +14,7 @@ public interface ProductRepository extends CrudRepository<Product, Integer> {
     
     Iterable<Product> findByUseBeforeLessThan(int useBefore);
     
-    Iterable<Product> findByPriceGreaterThan(int price);
+    Iterable<Product> findByPriceGreaterThan(BigDecimal price);
     
     @Query("select p from #{#entityName} p where p.manufactured = ?1 and p.useBefore = ?2")
     Iterable<Product> findByBestBeforeDate(int manufactured, int useBefore);
